pyproject.toml
[tool.poetry]
name = "mqtt-to-aprs"
version = "0.1.0"
description = "Send events from MQTT to APRS, via TCP and RF"
authors = ["Andrew Herrington <me@kg5jnc.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "mqtt_to_aprs"}]

[tool.poetry.scripts]
mqtt2aprs = "mqtt_to_aprs.cli:cli"

[tool.poetry.dependencies]
python = "^3.12"
aiomqtt = "^2.0.0"
pydantic = "^2.5.3"
aiofiles = "^23.2.1"
asyncstdlib = "^3.12.0"
asyncclick = "^8.1.7.1"
aprs3 = "^8.0.0.post1"
jmespath = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pre-commit = "^3.6.0"
ruff = "^0.1.14"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
bandit = "^1.7.6"
pre-commit-hooks = "^4.5.0"
reorder-python-imports = "^3.12.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
exclude_dirs = ["tests", ".github"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.mypy]
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.pytest.ini_options]
norecursedirs = ".github ci .git .idea"
addopts = "--cov=mqtt_to_aprs --cov-report xml:.coverage.xml --cov-report=term-missing"
